- Fix wrong variable set for SuperWeaponType `PreDependent` - Grouping some of BuildingClass `Mission::Missile` for debugging - Rname `Forshielded` to `ProtectType` and from `int` to `ProtectTypes` for better  reading - Fix abductor changing passanger owner to `Special` - Techno Flashing when receive damage additional - Some additional static variable for storing new locomotor(s) data - Add wide string version of locomotor string datas - Add `PhobosGlobal` class - Move some static variable to `PhobosGlobal` - Add `TechnoTypeExt` for `InfantryTypeExt` - Add additional `BulletTypeExt::CreateBullet`

+++ src/New/Interfaces/TestLocomotionClass.cpp
@@ -99,8 +99,8 @@ bool TestLocomotionClass::Process()
 		 *  Rotate the object around the center coord.
 		 */
 		int radius = Unsorted::LeptonsPerCell * 2;
-		coord.X += radius * std::sin(Angle);
-		coord.Y += radius * std::cos(Angle);
+		coord.X += int(radius * std::sin(Angle));
+		coord.Y += int(radius * std::cos(Angle));
 		//coord.Z // No need to adjust the height of the object.
 
 		// Pickup the object the game world before we set the new coord.
